Cleaning controller Composer data updated
